Add a title/comment to the pdb test files
This adds a title/comment to the pdb test files, so that they are unit tested
in test_io.py. For now the pdb->xyz conversion tests ignore this comment so
those tests are still running fine. This will be changed in a future commit.

While testing, it turned out that the reading of the comment in the pdb file
was reading the wrong line. This is also fixed with this commit.