even faster decorator
I investigated the possibility to include the start and stop calls
directly into the function at runtime. This is actually possible
(but not easy) reprogramming directly the bytecode. The problem is that
we should keep track of the reprogrammed functions in a dictionary.
This would lead to the same kind of behavior we have now:
check if the func_code of the function is already in a dictionary. The
only gain would be to avoid the call of a function that probably do not
make a so big difference.